Added command line arguments / help & removed old useless parser
