--- src/new-parser/grammar.ne
@@ -1,3 +1,0 @@
-@include "blocks.ne"
-
-main -> block {% id %}s
