+++ src/parser/grammar/grammar.ne
@@ -1,0 +1,3 @@
+@include "blocks.ne"
+
+main -> block {% id %}
