+++ tests/function.mcfunction
@@ -1,9 +1,1 @@
-year = 2018
-age = 19
-name = "Random Name"
-
-say Hello {{ "age + 1  \}}" }}!
-say You were born in {{ year - age }} because in {{ year }} you are {{ age }} years old!
-
-angle = PI/2
-say sin({{ angle }}) = {{ sin(angle) }}
+undefined
