+++ tests/function.mcscript
@@ -1,15 +1,1 @@
-{% if true %}
-variable = 5
-say    Hello World :D
-say {{ variable }}
-
-{% if false %}
-    say YES IT IS
-{% endif %}
-
-{% while variable > 0 %}
-    variable = variable - 2
-    say looping!
-{% endwhile %}
-
-{% endif %}
+undefined
